pyproject.toml
[project]
name = "bbva"
version = "0.1.0"
description = "BBVA Demo Access Code API"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "langchain>=0.0.335",
    "langgraph>=0.0.26",
    "openai>=1.3.0",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
    "redis>=5.0.1",
    "python-jose[cryptography]>=3.3.0",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
